third_party/blink/web_tests/platform/generic/fast/events/touch/create-touch-event-expected.txt

This tests whether the DOM can create TouchEvents.
PASS event is non-null.
PASS successfullyParsed is true

TEST COMPLETE
